requirements.txt
# Web server
flask
flask_sqlalchemy
gunicorn
# Spider
arrow
bs4
requests
pixivpy
